requirements.txt
Flask==0.10.1
Flask-DebugToolbar==0.10.0
Flask-SQLAlchemy==2.0
Gutenberg==0.4.0
Jinja2==2.7.3
MarkupSafe==0.23
SPARQLWrapper==1.6.4
SQLAlchemy==1.0.4
Werkzeug==0.10.4
argparse==1.2.1
blinker==1.3
html5lib==0.99999
isodate==0.5.1
itsdangerous==0.24
pyparsing==2.0.3
python-amazon-product-api==0.2.8
rdflib==4.2.0
requests==2.7.0
six==1.9.0
wsgiref==0.1.2
